Clear IPV6_V6ONLY.
Debian at least defaults to setting IPV6_V6ONLY to 1. That has a bunch
of advantages (e.g., you can use getaddrinfo to get all the addreses to
listen on), but this old code only is set up for one socket, and so its
easier to just clear the option for its socket.